tree/backend/requirements.txt
cmake==3.20.2
face-recognition==1.3.0
Pillow==7.0.0
